# mrd/openapi/communication-gateway-api.yaml
openapi: 3.0.3
info:
  title: Communication Gateway API
  description: |
    Communication Gateway for MEDrecord platform.
    
    **IMPORTANT: This is a SEPARATE Vercel project** from the main MedPlum repository.
    
    This gateway handles multi-channel patient communication:
    - WhatsApp (Twilio)
    - SMS (Twilio)
    - Email (Postmark)
    - Voice (ElevenLabs)
    
    The gateway receives webhooks from MedPlum subscriptions and delivers messages
    to patients via the configured channel.
  version: 1.0.0
  contact:
    name: MEDrecord Development Team
    url: https://medrecord.nl
servers:
  - url: https://comms.medrecord.nl
    description: Production Communication Gateway
...
paths:
...
  /api/callbacks/twilio/status:
    post:
      tags: [Callbacks]
      summary: Twilio message status callback
      description: |
        Receives delivery status updates from Twilio for WhatsApp and SMS messages.
      security: []
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/TwilioStatusCallback'
      responses:
        '200':
          description: Status processed
...
components:
  securitySchemes:
    SubscriptionSecret:
      type: apiKey
      in: header
      name: X-Subscription-Secret
      description: HMAC signature from MedPlum subscription

    PostmarkWebhook:
      type: apiKey
      in: header
      name: X-Postmark-Signature
      description: Postmark webhook signature

    ElevenLabsWebhook:
      type: apiKey
      in: header
      name: X-ElevenLabs-Signature
      description: ElevenLabs webhook signature
  schemas:
...
    TwilioStatusCallback:
      type: object
      properties:
        MessageSid:
          type: string
        MessageStatus:
          type: string
          enum: [queued, sent, delivered, undelivered, failed, read]
        To:
          type: string
        From:
          type: string
        ErrorCode:
          type: string
        ErrorMessage:
          type: string
...